+++ tests/test_node.py
@@ -2,65 +2,72 @@ import random
 import hashlib
 from collections.abc import Iterable
 
-from liaa.node import Node, NodeHeap
-from liaa.utils import hex_to_int
+from liaa.node import Node, NodeHeap, NodeType
+from liaa.utils import hex_to_int, pack
 
 
 class TestNode:
 	# pylint: disable=no-self-use
-	def test_node_long_id_derivation_is_ok(self):
-		rid = hashlib.sha1(str(random.getrandbits(255)).encode()).digest()
-		node = Node(rid)
-		assert node.long_id == hex_to_int(rid.hex())
+	def test_node_instance_attributes(self):
+		node = Node(key="127.0.0.1:8080", node_type=NodeType.Peer)
 
-	def test_valid_diff_calculation_is_ok(self):
-		ridone = hashlib.sha1(str(random.getrandbits(255)).encode())
-		ridtwo = hashlib.sha1(str(random.getrandbits(255)).encode())
+		assert node.ip == "127.0.0.1"
+		assert node.port == 8080
+		assert node.node_type == NodeType.Peer
+		assert not node.value
 
-		shouldbe = hex_to_int(ridone.digest().hex()) ^ hex_to_int(ridtwo.digest().hex())
-		none = Node(ridone.digest())
-		ntwo = Node(ridtwo.digest())
+		assert isinstance(node.long_id, int)
+		assert isinstance(node.digest, bytes)
+		assert node.digest == pack("I", node.key)
+		assert node.long_id < 2**160
+
+		assert str(node) == "peer@127.0.0.1:8080"
+
+	def test_distance_calculation(self, mkpeer):
+
+		addr1 = "127.0.0.1:8000"
+		addr2 = "127.0.0.1:9000"
+
+		none = mkpeer(key=addr1)
+		ntwo = mkpeer(key=addr2)
+
+		shouldbe = hex_to_int(pack("I", addr1).hex()) ^ hex_to_int(pack("I", addr2).hex())
+
 		assert none.distance_to(ntwo) == shouldbe
 
-	def test_distance_diff_of_same_node_is_0(self, mknode):
-		node_one = mknode(intid=1)
-		node_two = mknode(intid=1)
+	def test_node_is_same_node(self, mkpeer):
+		node_one = node_two = mkpeer()
 		assert node_one.distance_to(node_two) == 0
-
-	def test_is_same_node(self, mknode):
-		node_one = mknode(ip="0.0.0.0", port=0000)
-		node_two = mknode(ip="0.0.0.0", port=0000)
 		assert node_one.is_same_node(node_two)
 
-	def test_node_iter(self, mknode):
-		node = mknode()
-		assert tuple(node) == (node.digest_id, node.ip, node.port)
+	def test_node_iter(self, mkpeer):
+		node = mkpeer()
+		assert tuple(node) == (node.digest, node.ip, node.port)
 
 
 class TestNodeHeap:
 	# pylint: disable=no-self-use
-	def test_can_create_nodeheap(self, mknode):
-		heap = NodeHeap(mknode(), 2)
+	def test_can_create_nodeheap(self, mkpeer):
+		heap = NodeHeap(mkpeer(), 2)
 		assert isinstance(heap, NodeHeap)
 
-	def test_get_node_return_node_when_node_present(self, mknode):
-		heap = NodeHeap(mknode(), 3)
-		nodes = [mknode(intid=i) for i in range(3)]
-		# pylint: disable=invalid-name
-		for n in nodes:
-			heap.push(n)
-		node = heap.get_node(nodes[0].digest_id)
+	def test_get_node_return_node_when_node_present(self, mkpeer):
+		heap = NodeHeap(mkpeer(), 3)
+		nodes = [mkpeer() for _ in range(3)]
+		for node in nodes:
+			heap.push(node)
+		node = heap.get_node(nodes[0].digest)
 		assert isinstance(node, Node)
 
-	def test_get_node_returns_none_when_node_not_exists(self, mknode):
-		heap = NodeHeap(mknode(), 1)
+	def test_get_node_returns_none_when_node_not_exists(self, mkpeer):
+		heap = NodeHeap(mkpeer(), 1)
 		empty = heap.get_node(123)
 		assert not empty
 
-	def test_mark_contacted_works_ok(self, mknode):
+	def test_mark_contacted_works_ok(self, mkpeer):
 		maxsize = 10
-		heap = NodeHeap(mknode(), maxsize)
-		nodes = [mknode(intid=i) for i in range(maxsize)]
+		heap = NodeHeap(mkpeer(), maxsize)
+		nodes = [mkpeer() for _ in range(maxsize)]
 		for node in nodes:
 			heap.push(node)
 		contacted = nodes[:5]
@@ -71,64 +78,58 @@ class TestNodeHeap:
 		assert not heap.have_contacted_all()
 		assert len(heap.get_uncontacted()) == 5
 
-	def test_popleft_returns_left_if_heap_not_empty(self, mknode):
+	def test_popleft_returns_left_if_heap_not_empty(self, mkpeer):
 		maxsize = 5
-		heap = NodeHeap(mknode(), maxsize)
-		nodes = [mknode(intid=i) for i in range(maxsize)]
+		heap = NodeHeap(mkpeer(), maxsize)
+		nodes = [mkpeer() for _ in range(maxsize)]
 		for node in nodes:
 			heap.push(node)
 
 		popped = heap.popleft()
 		assert isinstance(popped, Node)
 
-	def test_popleft_returns_none_when_heap_empty(self, mknode):
+	def test_popleft_returns_none_when_heap_empty(self, mkpeer):
 		maxsize = 1
-		heap = NodeHeap(mknode(), maxsize)
-		nodes = [mknode(intid=i) for i in range(maxsize)]
+		heap = NodeHeap(mkpeer(), maxsize)
+		nodes = [mkpeer()]
 		for node in nodes:
 			heap.push(node)
 
 		heap.remove(nodes)
+
 		popped = heap.popleft()
-		assert isinstance(popped, Iterable)
-		# pylint: disable=invalid-name
-		_, ip, port = tuple(popped)
-		assert (not ip) and (not port)
+		assert not popped
 
-
-	def test_heap_overload_doesnt_exceed_maxsize(self, mknode):
+	def test_heap_overload_doesnt_exceed_maxsize(self, mkpeer):
 		maxsize = 3
-		node = NodeHeap(mknode(intid=0), maxsize)
+		node = NodeHeap(mkpeer(), maxsize)
 		assert not node
 
 		for digit in range(10):
-			node.push(mknode(intid=digit))
+			node.push(mkpeer())
 
 		assert len(node) == maxsize
 		assert len(list(node)) == maxsize
 
-	def test_heap_iters_over_nsmallest_via_distance(self, mknode):
-		heap = NodeHeap(mknode(intid=0), 5)
-		nodes = [mknode(intid=x) for x in range(10)]
-		for index, node in enumerate(nodes):
-			heap.push(node)
-
-		for index, node in enumerate(heap):
-			assert index == node.long_id
-			assert index < 5
-
-	def test_remove(self, mknode):
-		maxsize = 5
-		heap = NodeHeap(mknode(intid=0), maxsize)
-		nodes = [mknode(intid=x) for x in range(10)]
+	def test_heap_iters_over_nsmallest_via_distance(self, mkpeer):
+		heap = NodeHeap(mkpeer(), 5)
+		nodes = [mkpeer() for _ in range(10)]
 		for node in nodes:
 			heap.push(node)
 
-		heap.remove([nodes[0].digest_id, nodes[1].digest_id])
-		# len(heap) == min(10-2, maxsize)
+		for index, node in enumerate(heap):
+			assert index < 5
+
+	def test_remove(self, mkpeer):
+		maxsize = 5
+		heap = NodeHeap(mkpeer(), maxsize)
+		nodes = [mkpeer() for _ in range(10)]
+		for node in nodes:
+			heap.push(node)
+
+		heap.remove([nodes[0], nodes[1]])
 		assert len(list(heap)) == maxsize
 
 		for index, node in enumerate(heap):
 			# we removed to elements so offset index to account for it
-			assert index + 2 == node.long_id
 			assert index < maxsize