+++ docs/CHANGELOG.md
@@ -24,3 +24,8 @@
    - [refactor] remove config
    - [enhancement] add Docker support
    - [chore] rename project to `liaa` to avoid pypi namespace collisions
+   - [bugfix] ensure crawler working properly (ISSUE-28)
+      - remove sphinx docs and add new docs
+      - update CI to push on version branches and master
+      - base node IDs off of ip:port, and search keys and adjust tests
+